feat(routing): add authentication route handlers and diagnostics
- Introduced auth.route-diagnostics.server.ts for logging and diagnostics of authentication routes.
- Created auth.route-handlers.account.server.ts to define account-related authentication handlers.
- Added auth.route-handlers.instances.server.ts for instance management authentication handlers.
- Defined auth.route-handlers.types.ts to establish types for authentication handlers and routes.
- Implemented auth.route-runtime.server.ts to manage route resolution and method not allowed responses.
